opraven vyber pozice spawnu itemu, zkracen spawnovaci interval

+++ itemspawn.py
@@ -33,13 +33,13 @@ class ItemSpawner:
         side = random.choice(["t", "b", "l", "r"])
         # top
         if side == "t":
-            return np.array([random.randint(0, ScreenSetup.width), random.randint(0, 1/5*ScreenSetup.height)])
+            return np.array([random.randint(0, ScreenSetup.width), random.randint(0, int(1/5*ScreenSetup.height))])
         # bottom
         elif side == "b":
-            return np.array([random.randint(0, ScreenSetup.width), random.randint(4/5*ScreenSetup.height, ScreenSetup.height)])
+            return np.array([random.randint(0, ScreenSetup.width), random.randint(int(4/5*ScreenSetup.height), ScreenSetup.height)])
         # left
         elif side == "l":
-            return np.array([random.randint(0, 1/6*ScreenSetup.width), random.randint(0, ScreenSetup.height)])
+            return np.array([random.randint(0, int(1/6*ScreenSetup.width)), random.randint(0, ScreenSetup.height)])
         # right
         else:
-            return np.array([random.randint(5/6*ScreenSetup.width, ScreenSetup.width), random.randint(0, ScreenSetup.height)])
+            return np.array([random.randint(int(5/6*ScreenSetup.width), ScreenSetup.width), random.randint(0, ScreenSetup.height)])
